+++ packages/@webex/plugin-meetings/src/meeting/util.ts
@@ -12,7 +12,9 @@ import {
   FULL_STATE,
   SELF_POLICY,
   EVENT_TRIGGERS,
+  IP_VERSION,
 } from '../constants';
+import BrowserDetection from '../common/browser-detection';
 import IntentToJoinError from '../common/errors/intent-to-join';
 import JoinMeetingError from '../common/errors/join-meeting';
 import ParameterError from '../common/errors/parameter';
@@ -72,7 +74,6 @@ const MeetingUtil = {
           audioMuted,
           videoMuted,
         },
-        ipVersion: meeting.getWebexObject().meetings.reachability.getIpVersion(),
       })
       .then((response) => {
         // @ts-ignore
@@ -90,6 +91,38 @@ const MeetingUtil = {
   isOwnerSelf: (owner, selfId) => owner === selfId,
 
   isPinOrGuest: (err) => err?.body?.errorCode && INTENT_TO_JOIN.includes(err.body.errorCode),
+
+  /**
+   * Returns the current state of knowledge about whether we are on an ipv4-only or ipv6-only or mixed (ipv4 and ipv6) network.
+   * The return value matches the possible values of "ipver" parameter used by the backend APIs.
+   *
+   * @param {Object} webex webex instance
+   * @returns {IP_VERSION|undefined} ipver value to be passed to the backend APIs or undefined if we should not pass any value to the backend
+   */
+  getIpVersion(webex: any): IP_VERSION | undefined {
+    const {supportsIpV4, supportsIpV6} = webex.internal.device.ipNetworkDetector;
+
+    if (BrowserDetection().isBrowser('firefox')) {
+      // our ipv6 solution relies on FQDN ICE candidates, but Firefox doesn't support them,
+      // see https://bugzilla.mozilla.org/show_bug.cgi?id=1713128
+      // so for Firefox we don't want the backend to activate the "ipv6 feature"
+      return undefined;
+    }
+
+    if (supportsIpV4 && supportsIpV6) {
+      return IP_VERSION.ipv4_and_ipv6;
+    }
+
+    if (supportsIpV4) {
+      return IP_VERSION.only_ipv4;
+    }
+
+    if (supportsIpV6) {
+      return IP_VERSION.only_ipv6;
+    }
+
+    return IP_VERSION.unknown;
+  },
 
   joinMeeting: (meeting, options) => {
     if (!meeting) {
@@ -112,6 +145,7 @@ const MeetingUtil = {
         meetingNumber: meeting.meetingNumber,
         deviceUrl: meeting.deviceUrl,
         locusUrl: meeting.locusUrl,
+        locusClusterUrl: meeting.meetingInfo?.locusClusterUrl,
         correlationId: meeting.correlationId,
         roapMessage: options.roapMessage,
         permissionToken: meeting.permissionToken,
@@ -125,7 +159,7 @@ const MeetingUtil = {
         locale: options.locale,
         deviceCapabilities: options.deviceCapabilities,
         liveAnnotationSupported: options.liveAnnotationSupported,
-        ipVersion: meeting.getWebexObject().meetings.reachability.getIpVersion(),
+        ipVersion: MeetingUtil.getIpVersion(meeting.getWebexObject()),
       })
       .then((res) => {
         // @ts-ignore