+++ packages/@webex/plugin-meetings/src/roap/turnDiscovery.ts
@@ -8,6 +8,7 @@ import {ROAP} from '../constants';
 
 import RoapRequest from './request';
 import Meeting from '../meeting';
+import MeetingUtil from '../meeting/util';
 
 const TURN_DISCOVERY_TIMEOUT = 10; // in seconds
 
@@ -183,7 +184,7 @@ export default class TurnDiscovery {
         meetingId: meeting.id,
         locusMediaRequest: meeting.locusMediaRequest,
         // @ts-ignore - because of meeting.webex
-        ipVersion: meeting.webex.meetings.reachability.getIpVersion(),
+        ipVersion: MeetingUtil.getIpVersion(meeting.webex),
       })
       .then(({mediaConnections}) => {
         if (mediaConnections) {
@@ -214,8 +215,6 @@ export default class TurnDiscovery {
       mediaId: meeting.mediaId,
       meetingId: meeting.id,
       locusMediaRequest: meeting.locusMediaRequest,
-      // @ts-ignore - because of meeting.webex
-      ipVersion: meeting.webex.meetings.reachability.getIpVersion(),
     });
   }